refactor: 이미지 webp 형식으로 변경 #227
